requirements.txt
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
